Fix a couple more signed/unsigned warnings.
I still don't know why these only appear on Jenkins...

--HG--
branch : release-2.x.y
extra : amend_source : 9b2f58758d0a6b8d2b7a664705529483efa2d057

+++ libpurple/dnsquery.c
@@ -421,7 +421,7 @@ cope_with_gdb_brokenness(void)
 	if(n < 0)
 		return;
 
-	e[MIN(n,sizeof(e)-1)] = '\0';
+	e[MIN((gsize)n,sizeof(e)-1)] = '\0';
 
 	if(strstr(e,"gdb")) {
 		purple_debug_info("dns",
